move handleToolCallValidation out of thread service
